requirements.txt
pyttsx3==2.90
wikipedia==1.4.0
SpeechRecognition==3.10.0
AppOpener==1.6
openai==1.30.1
python-dotenv==1.0.1